# Reduce thread contention, allowing 8-12% faster jackhmmer