@Augustin-Zidek 2024-11-14
I don't think IO contention or FASTA parsing is the reason, because:

1. We typically run jackhmmer/nhmmer programs against databases that are stored in a RAM-disk (`/dev/shm/...`), so we have very fast IO on these. 
2. If we shard the in-RAM FASTA *k*-ways on a single machine and run *k* jackhmmer/nhmmer instances against these shards in parallel, we get basically linear scaling with *k* up to a certain number (somewhere around 64-128, depending on the machine). This strongly suggests that we are not saturating IO.
3. Our C++ FASTA parser can read and parse the in-RAM db much faster even with a single thread. Of course, this is just paring the FASTA, not doing any of the heavy lifting that jackhmmer does, but it shows the theoretical lower bound on the achievable speed.
4. Our observations are consistent with PyHMMER which also scales nicely with the number of threads: https://pyhmmer.readthedocs.io/en/stable/guide/benchmarks.html