@Augustin-Zidek 2024-11-18
> Huh, those pyhmmer results are really interesting. I haven't had a chance to look at HMMER's scaling with core count in detail, but it's the next big thing on my to-do list after my current project (porting HMMER to use the AVX and AVX-512 SIMD ISAs). I'll definitely be looking at pyhmmer to see what it's doing differently.

AVX and AVX-512 sound very exciting, can't wait (and thank you!).

> Is your FASTA parser available somewhere that I could take a look at it?

Indeed, here it is: 

* https://github.com/google-deepmind/alphafold3/blob/main/src/alphafold3/parsers/cpp/fasta_iterator_lib.h
* https://github.com/google-deepmind/alphafold3/blob/main/src/alphafold3/parsers/cpp/fasta_iterator_lib.cc

But I am guessing you parser does significantly more than we do, like encoding each amino acid as a numeric value.