@npcarter 2024-11-18
Quick update:  I took a look at the pyhmmer results you linked to, and they're avoiding the file parsing bottleneck by caching the parsed sequence/HMM file in memory so that they only have to parse it once.  That works very well for the somewhat-unusual (at least, in terms of how we run HMMER in our lab) benchmark they're running, which is searching a 2,100-sequence database against the ~20K Pfam HMMs.  For something like that, caching the parsed database will give you big speedups.  For the case we see more commonly, where someone is searching one or a few query objects against a large database, it won't help that much, and may well fail completely by overloading the memory of your machine.

If your use case involves many searches of the same database, we have a new version of the HMMER server that's in beta and will do sharded caching of parsed databases in RAM.  It gets very good speedups out to at least 500 cores used per search, and we could really use some people to help stress-test it.  If that's of interest to you, let's talk.