@Augustin-Zidek 2024-11-22
Yes, one of our use-cases (alphafoldserver.com) involves searching the same db with different inputs. However, we run the searches one-by-one (as user queries arrive), rather than in bulk. However, being able to cache the parsed db in RAM would be very nice for this use-case.

Our other use-case is within AlphaFold:

* https://github.com/google-deepmind/alphafold3/blob/main/src/alphafold3/data/tools/jackhmmer.py
* https://github.com/google-deepmind/alphafold3/blob/main/src/alphafold3/data/tools/nhmmer.py
* https://github.com/google-deepmind/alphafold3/blob/main/src/alphafold3/data/pipeline.py#L30

In that case we again have the same set of fixed dbs, and it would be ideal to have as low latency as possible.

However, how do you feel about this change? Are you willing to merge it?