@npcarter 2024-11-25
Just got done running some tests, and this change looks good, so I'm going to accept it.